+++ Fun.Build/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## [Unreleased]
 
+## [1.1.7] - 2024-09-11
+
+- Add overloads for **when' (stage "some condition" {})**
+- Add helper CE builder **whenStage**
+
 ## [1.1.6] - 2024-07-22
 
 - Add RunSensitiveCommandCaptureOutput